backend/tag-service/requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
pymongo==4.6.0
python-dotenv==1.0.0
PyJWT==2.8.0
cryptography==41.0.7
